Rework the model type configuration and evaluation.
This PR changes the configuration file from pure JSON to a joint JSON and Parquet for smaller size on disk and simpler handling within the code base.

Furthermore, the pathway dependent function application has been reworked to drop mirroring the old JSON structure and take full advantage of the speed of numpy. In experimental tests, this gives roughly a 40× speedup.

Addresses the consuming side of BBPP134-284